refactor: decouple project index file discovery
